api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi[binaries]>=25.1.0",
    "asgi-correlation-id>=4.3.4",
    "basemodel",
    "fastapi[standard]>=0.121.2",
    "gdown>=5.2.0",
    "numpy>=2.3.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.44",
    "structlog>=25.5.0",
    "ulid-py>=1.1.0",
]
